four_hours/requirements.txt
asgiref==3.6.0
autopep8==2.0.0
black==23.1.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
firebase-admin==6.1.0
firebase-auth==3.0.28
flex==6.14.1
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpointer==2.3
jws==0.1.3
MarkupSafe==2.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
oauth2client==3.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3987==1.3.8
rsa==4.9
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
strict-rfc3339==0.7
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
validate-email==1.3
